housing_price_api/requirements.txt
asgiref==3.8.0
Django==4.2.11
django-cors-headers==4.3.1
django-rest==0.8.7
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
joblib==1.2.0
numpy==1.23.4
packaging==24.0
pandas==1.5.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.10.0
uritemplate==4.1.1
